move tests to gatsby folder so its more organized
